try refactor network
