--- src/store/models/models.types.ts
@@ -1,7 +1,0 @@
-export type GetSetType<T> = (
-  partial: T | Partial<T> | ((state: T) => void),
-  shouldReplace?: boolean | undefined,
-  type?: {
-    type: string;
-  }
-) => void;
