+++ src/store/models/network/helpers.ts
@@ -1,0 +1,26 @@
+import { CurrentNetworkType } from '../../../types/network.types';
+
+export const defaultNetwork: CurrentNetworkType = {
+  id: 'not-configured',
+  chainId: '',
+  name: 'NOT CONFIGURED',
+  egldLabel: '',
+  decimals: '18',
+  digits: '4',
+  gasPerDataByte: '1500',
+  walletConnectDeepLink: '',
+  walletConnectBridgeAddress: '',
+  walletConnectV2RelayAddress: '',
+  walletConnectV2ProjectId: '',
+  walletConnectV2Options: {},
+  walletAddress: '',
+  apiAddress: '',
+  explorerAddress: '',
+  apiTimeout: '4000'
+};
+
+export function getRandomAddressFromNetwork(walletConnectAddresses: string[]) {
+  return walletConnectAddresses[
+    Math.floor(Math.random() * walletConnectAddresses.length)
+  ];
+}
